move fake class
