+++ core/data/src/main/java/dev/dai/android/architecture/core/data/user/fake/FakeUserRepository.kt
@@ -1,4 +1,4 @@
-package dev.dai.android.architecture.template.core.testing.repository
+package dev.dai.android.architecture.core.data.user.fake
 
 import dev.dai.android.architecture.core.data.user.UserRepository
 import dev.dai.android.architecture.template.core.model.User
